\name{etable-package}
\alias{etable-package}
\alias{etable}
\docType{package}
\title{
Easy Table
}
\description{
The package comes without any warranty.
}
\details{
\tabular{ll}{
Package: \tab etable\cr
Title: \tab Easy Table\cr
Type: \tab Package\cr
Version: \tab 1.3.0\cr
Date: \tab 2021-05-23\cr
Depends: \tab R (>= 3.0.0)\cr
Imports: \tab xtable, Hmisc\cr
License: \tab GPL version 3 or newer\cr
LazyLoad: \tab yes\cr
}
}
\author{
Andreas Schulz\cr
Maintainer: <ades-s@web.de>
}
