\name{etable-package}
\alias{etable-package}
\alias{etable}
\docType{package}
\title{
Easy Table
}
\description{
The package comes without any warranty.
}
\details{
\tabular{ll}{
Package: \tab etable\cr
Title: \tab Easy Table\cr
Type: \tab Package\cr
Version: \tab 1.0.3\cr
Date: \tab 2011-12-19\cr
Depends: \tab epade, Hmisc, R (>= 2.13)\cr
License: \tab GPL version 2 or newer\cr
LazyLoad: \tab yes\cr
}
}
\author{
Andreas Schulz\cr
Maintainer: <ades-s@web.de>
}
